requirements.txt
chardet
fastapi
fire
numpy
pandas
uvicorn
uuid
weaviate-client
